backend/requirement.txt
fastapi 
uvicorn 
requests 
python-dotenv 
pydantic
beautifulsoup4
httpx
apscheduler
sqlalchemy